+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export {usePrefersReducedMotion} from './usePrefersReducedMotion';
 export {useLinearXAxisDetails, ChartDetails} from './useLinearXAxisDetails';
 export {useLinearXScale} from './useLinearXScale';
+export {usePrevious} from './use-previous';